CLN move thread tests into test_thread_executor
